* 'master' of git://github.com/Seldaek/slippy:
  Add support for data-background for download button, fixes #12
  Fix bug with adjacent slides containing backrgounds
  Adjust overview style
  Added support for full-screen images per slide
  Clean up of index.php
  fixing issues mentioned by sealdek and making web download work again
  tweak image handling: baseWidth: false to disable, noscale to disable single image, add imgScaleTrivial to avoid mini-scaling
  adding cli handling to index.php and explain download in README
  Support spaces in input for pdf conversion
  Updated README
  Speed up PDF rendering by using multiple concurrent renderers
  Update PhantomJS script to 1.3.0
  Remove TAB handler to show the overview (quite annoying when alt-tabbing through apps)
  Fix js execute button to work with html entities in code block
  Some fixes to the PDF rendering

+++ bin/slippy-pdf.sh
@@ -7,6 +7,7 @@ if [ "" = "$2" ]; then
     echo "Usage: ./phantom-pdf.sh <slides URL> <PDF filename>"
     exit 1
 fi
+
 mkdir -p tmp-pdf/
 rm -f tmp-pdf/*.png
 rm -f tmp-pdf/*.pdf
@@ -35,7 +36,12 @@ if [ "" = "$pdftk" ]; then
     fi
 fi
 
-$phantom $bin/phantom-slippy-to-pdf.js $1 tmp-pdf/
+$phantom $bin/phantom-slippy-to-pdf.js "$1" tmp-pdf/
+if [ "$?" != "0" ]
+then
+    echo 'PhantomJS error, aborting.'
+    exit
+fi
 
-$pdftk tmp-pdf/*.pdf cat output $2
+$pdftk tmp-pdf/*.pdf cat output "$2"
 rm -r tmp-pdf/
